Relocate embed-mode.js to comply with dependency limitations
