+++ src/amp-story-player/amp-story-player-impl.js
@@ -15,10 +15,7 @@
  */
 
 import * as ampToolboxCacheUrl from '@ampproject/toolbox-cache-url';
-import {
-  EmbedMode,
-  EmbedModeParam,
-} from '../../extensions/amp-story/1.0/embed-mode';
+import {EmbedMode, EmbedModeParam} from '../embed-mode';
 import {IframePool} from './amp-story-player-iframe-pool';
 import {Messaging} from '@ampproject/viewer-messaging';
 import {VisibilityState} from '../visibility-state';
